requirements-dev.txt
playwright>=1.45.0
pillow>=10.0.0
